bundle move to swaggerknife
