+++ examples/library/walkers00.output
@@ -1,3 +1,4 @@
+definitions/color/$ref #/definitions/thing/properties/colors/black-as-the-night
 definitions/person/$ref schemas/people/Bruce-Wayne.json
 definitions/place/$ref schemas/places.yaml#/definitions/Gotham-City
 definitions/thing/$ref http://wayne-enterprises.com/things/batmobile
